requirements.txt
streamlit>=1.33.0
pymupdf>=1.24.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
huggingface_hub>=0.23.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
